[Core] Added UIMenuBar's bindings to FrameworkLuaLib; [Core] Added LuaEventGroup; [Core] Added UIPanel Properties Lua Table; [Core] Added UIPanel Event Bindings to FrameworkLuaLib; [Core] Made SoundManager reuse Sounds and Musics; [Core] Added Cleanup to SoundManager; [Core] Added Lua On Item Selected Callback for UIMenuBar;

+++ Development/Core/Sprite.cpp
@@ -130,9 +130,9 @@ namespace FlamingTorch
 		{
 			//1st: Normalize
 			{
-				Vector2 NormalizedSize = Vector2(SpriteTexture->GetIndex().Width() - 2, SpriteTexture->GetIndex().Height() - 2) / SpriteTexture->GetIndex().Owner->MainTexture->Size();
-				Vector2 NormalizedPosition = Vector2(SpriteTexture->GetIndex().Owner->Indices[SpriteTexture->GetIndex().Index].x + 1, SpriteTexture->GetIndex().Owner->Indices[SpriteTexture->GetIndex().Index].y + 1) /
-					SpriteTexture->GetIndex().Owner->MainTexture->Size();
+				Vector2 NormalizedSize = Vector2(SpriteTexture->GetIndex().Width() - 2.0f, SpriteTexture->GetIndex().Height() - 2.0f) / SpriteTexture->GetIndex().Owner->MainTexture->Size();
+				Vector2 NormalizedPosition = Vector2((f32)SpriteTexture->GetIndex().Owner->Indices[SpriteTexture->GetIndex().Index].x + 1,
+					(f32)SpriteTexture->GetIndex().Owner->Indices[SpriteTexture->GetIndex().Index].y + 1) / SpriteTexture->GetIndex().Owner->MainTexture->Size();
 
 				//2nd: Multiply by size, then translate
 				for(uint32 i = 0; i < 6; i++)
